Issue #171: update QEverCloud to 6.2.0 and update API calls to it across the codebase

+++ src/threads/syncrunner.cpp
@@ -1229,7 +1229,7 @@ void SyncRunner::communicationErrorHandler() {
         apiRateLimitExceeded = true;
     }
     if (type == CommunicationError::EDAMUserException) {
-        if (comm->getLastErrorCode() == EDAMErrorCode::AUTH_EXPIRED) {
+        if (comm->getLastErrorCode() == static_cast<int>(EDAMErrorCode::AUTH_EXPIRED)) {
             global.accountsManager->setOAuthToken("");
         }
     }
